+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Fix potential conflict causing Gutenberg blocks not to load
+* Simplify clearing the Recras cache
 
 ## 4.0.2 (2020-08-18)
 * Fix max length of various contact form fields